services/extract_topic_service/requirements.txt
torch
transformers
tokenizers
sentence-transformers
huggingface_hub
flashtext
psycopg2-binary
py_vncorenlp
underthesea
confluent-kafka